fixed cost counting. added more detail to log

+++ modules/worker.h
@@ -70,7 +70,7 @@ class Worker {
   long getInstructionsPerTime();
   long getTotalExecutionTime();
   long getTotalCPUTime();
-  long getTotalCost();
+  float getTotalCost();
   int getQueuedJobs();
   bool ping();
